Small Tweaks
Lots of small stuff. Moving pre-existing rope down a little bit so it does clip out of the box, moving banners towards walls more in some cases. Adding more things to room 10, since it seemed barren. Added a bunch of objects into the wall of the object its on for scoot.